implementing contact list
