+++ lib/main.dart
@@ -1,4 +1,5 @@
-import 'package:bytebank/screens/transferencia/lista.dart';
+import 'package:bytebank/screens/contacts_list.dart';
+import 'package:bytebank/screens/dashboard.dart';
 import 'package:flutter/material.dart';
 
 void main() => runApp(BytebankApp());
@@ -18,57 +19,6 @@ class BytebankApp extends StatelessWidget {
         ),
       ),
       home: Dashboard(),
-    );
-  }
-}
-
-class Dashboard extends StatelessWidget {
-  const Dashboard({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('Dashboard'),
-      ),
-      body: Column(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Image.asset('images/bytebank_logo.png'),
-          ),
-          Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Container(
-              padding: EdgeInsets.all(8.0),
-              height: 100,
-              width: 150,
-              color: Theme.of(context).primaryColor,
-              child: Column(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Icon(
-                    Icons.people,
-                    color: Colors.white,
-                    size: 24.0,
-                  ),
-                  Text(
-                    'Contacts',
-                    style: TextStyle(
-                      color: Colors.white,
-                      height: 1.0,
-                      fontSize: 16,
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          )
-        ],
-      ),
     );
   }
 }
