+++ lib/screens/contacts_list.dart
@@ -1,0 +1,34 @@
+import 'package:flutter/material.dart';
+
+class ContactsList extends StatelessWidget {
+  const ContactsList({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('Contacts'),
+      ),
+      body: ListView(
+        children: [
+          Card(
+            child: ListTile(
+              title: Text(
+                'Alex',
+                style: TextStyle(fontSize: 24.0),
+              ),
+              subtitle: Text(
+                '1000',
+                style: TextStyle(fontSize: 16.0),
+              ),
+            ),
+          ),
+        ],
+      ),
+      floatingActionButton: FloatingActionButton(
+        onPressed: () {},
+        child: Icon(Icons.add),
+      ),
+    );
+  }
+}
